Flytt ut faktisk trygdetid table i egne komponenter
